refactor: rename TeamPermission to UserPermission
